@@ -1,3 +1,9 @@
+/* eslint-disable @typescript-eslint/no-empty-function */
+import fs from "fs"
+import path from "path"
+import chalk from "chalk"
+import * as process from "process"
+
 import {
     Command,
     CommandArg,
@@ -5,228 +11,256 @@ import {
     matchName,
     Terminal,
     ToolController,
-} from "../core";
+} from "../core"
+import { controllers, findControllerAndCommandByAlias } from "../controllers"
+import { formatTable } from "../core/utils"
+import { printUsage } from "../everdev/help"
+import { printSummaryInfo } from "../everdev/info"
 import {
-    controllers,
-    findControllerAndCommandByAlias,
-} from "../controllers";
-import {
-    formatTable,
-} from "../core/utils";
-import { printUsage } from "../everdev/help";
-import { printSummaryInfo } from "../everdev/info";
-import { createLatestVerFile, getUpdateIsAvailableMsg } from "../everdev/checkNewVersion"; 
-import * as process from "process";
-import fs from "fs";
-import path from "path";
-import chalk from "chalk";
+    createLatestVerFile,
+    getUpdateIsAvailableMsg,
+} from "../everdev/checkNewVersion"
 
 function findOptionArg(command: Command, name: string): CommandArg | undefined {
     if (name.startsWith("--")) {
-        name = name.substr(2).toLowerCase();
-        return command.args?.find(x => !x.isArg && x.name === name);
+        name = name.substring(2).toLowerCase()
+        return command.args?.find(x => !x.isArg && x.name === name)
     }
     if (name.startsWith("-")) {
-        name = name.substr(1).toLowerCase();
-        return command.args?.find(x => !x.isArg && x.alias === name);
+        name = name.substring(1).toLowerCase()
+        return command.args?.find(x => !x.isArg && x.alias === name)
     }
-    return undefined;
+    return undefined
 }
 
+export class CommandLine {
+    args: Record<string, string | boolean> = {}
+    controller: ToolController | undefined = undefined
+    command: Command | undefined = undefined
+    positional: CommandArg[] = []
+    unresolved = new Map<string, CommandArg>()
+    pending: CommandArg | undefined = undefined
+    printSummaryInfo = false
+    greedyArgument: CommandArg
 
-class CommandLine {
-    args: { help?: boolean, [name: string]: any } = {};
-    controller: ToolController | undefined = undefined;
-    command: Command | undefined = undefined;
-    positional: CommandArg[] = [];
-    unresolved = new Map<string, CommandArg>();
-    pending: CommandArg | undefined = undefined;
-    printSummaryInfo = false;
-
-    setArgValue(arg: CommandArg, value: any) {
+    setArgValue(arg: CommandArg, value: string | boolean) {
         const name = arg.name
             .split("-")
-            .map((x, i) => i > 0 ? (x.substr(0, 1).toUpperCase() + x.substr(1)) : x)
-            .join("");
-        this.args[name] = value;
+            .map((x, i) =>
+                i > 0 ? x.substring(0, 1).toUpperCase() + x.substring(1) : x,
+            )
+            .join("")
+
+        if (typeof value === "string") {
+            const prevVal = this.args[name]
+            if (prevVal) {
+                this.args[name] = prevVal + " " + value
+            } else {
+                this.args[name] = value
+            }
+        } else {
+            this.args[name] = value
+        }
     }
 
-    async resolveValue(arg: CommandArg, value: string | undefined) {
+    async resolveValue(arg: CommandArg, value?: string) {
         if (arg.type === "boolean" && value === undefined) {
-            value = "true";
+            value = "true"
         }
+
         if (value === undefined) {
-            throw new Error(`Missing value for ${arg.name}`);
+            throw new Error(`Missing value for ${arg.name}`)
         }
-        let resolved: any;
+        let resolved: string | boolean
         if (arg.type === "boolean") {
-            resolved = value.toLowerCase() === "true";
+            resolved = value.toLowerCase() === "true"
         } else {
-            resolved = value;
+            resolved = value
         }
-        const variants = await getArgVariants(arg);
+        const variants = await getArgVariants(arg)
         if (variants && !variants.find(x => x.value === value)) {
-            throw await missingArgError(arg);
+            throw await missingArgError(arg)
         }
-        this.setArgValue(arg, resolved);
-        this.unresolved.delete(arg.name);
-        const i = this.positional.indexOf(arg);
+        this.setArgValue(arg, resolved)
+        this.unresolved.delete(arg.name)
+        const i = this.positional.indexOf(arg)
         if (i >= 0) {
-            this.positional.splice(i, 1);
+            this.positional.splice(i, 1)
         }
         if (this.pending === arg) {
-            this.pending = undefined;
+            this.pending = undefined
         }
     }
 
     async resolveDefault(arg: CommandArg) {
         if (arg.defaultValue !== undefined) {
-            this.setArgValue(arg, arg.type === "boolean" ? arg.defaultValue === "true" : arg.defaultValue);
+            this.setArgValue(
+                arg,
+                arg.type === "boolean"
+                    ? arg.defaultValue === "true"
+                    : arg.defaultValue,
+            )
         } else if (arg.type === "folder") {
-            this.setArgValue(arg, process.cwd());
+            this.setArgValue(arg, process.cwd())
         } else {
-            throw await missingArgError(arg);
+            throw await missingArgError(arg)
         }
     }
 
     async parseOptionName(name: string) {
         if (this.pending) {
-            await this.resolveValue(this.pending, undefined);
+            await this.resolveValue(this.pending, undefined)
         }
-        const optionName: string = name.toLowerCase();
+        const optionName: string = name.toLowerCase()
         if (optionName === "--help" || optionName === "-h") {
-            this.args.help = true;
+            this.args.help = true
         } else if (this.command) {
-            this.pending = findOptionArg(this.command, optionName);
+            this.pending = findOptionArg(this.command, optionName)
             if (!this.pending) {
-                throw new Error(`Unknown option ${optionName}`);
+                throw new Error(`Unknown option ${optionName}`)
             }
             if (this.pending.type === "boolean") {
-                await this.resolveValue(this.pending, undefined);
+                await this.resolveValue(this.pending, undefined)
             }
         } else if (this.controller) {
-            throw new Error(`Unexpected option ${optionName} before command name.`);
+            throw new Error(
+                `Unexpected option ${optionName} before command name.`,
+            )
         } else {
-            throw new Error(`Unexpected option ${optionName} before tool name.`);
+            throw new Error(`Unexpected option ${optionName} before tool name.`)
         }
     }
 
     setCommand(command: Command) {
-        this.command = command;
+        this.command = command
         for (const arg of this.command.args ?? []) {
-            this.unresolved.set(arg.name, arg);
+            this.unresolved.set(arg.name, arg)
             if (arg.isArg) {
-                this.positional.push(arg);
+                this.positional.push(arg)
             }
         }
-
     }
 
     async parse(programArgs: string[]) {
-        for (let arg of programArgs) {
+        for (let i = 0; i < programArgs.length; i++) {
+            let arg = programArgs[i]
             if (arg.startsWith("-") && !this.pending) {
-                await this.parseOptionName(arg);
+                await this.parseOptionName(arg)
             } else {
-                arg = arg.trim();
+                arg = arg.trim()
                 if (this.pending) {
-                    await this.resolveValue(this.pending, arg);
+                    await this.resolveValue(this.pending, arg)
                 } else if (this.controller && this.command) {
-                    if (this.positional.length === 0) {
-                        throw new Error(`Unexpected argument ${arg}`);
+                    if (
+                        this.positional.length === 0 &&
+                        this.greedyArgument === undefined
+                    ) {
+                        throw new Error(`Unexpected argument ${arg}`)
                     }
-                    await this.resolveValue(this.positional[0], arg);
+                    if (this.positional[0]?.greedy) {
+                        this.greedyArgument = this.positional[0]
+                    }
+                    await this.resolveValue(
+                        this.positional[0] || this.greedyArgument,
+                        arg,
+                    )
                 } else if (this.controller) {
-                    const command = this.controller.commands.find(x => matchName(x, arg));
+                    const command = this.controller.commands.find(x =>
+                        matchName(x, arg),
+                    )
                     if (command) {
-                        this.setCommand(command);
+                        this.setCommand(command)
                     } else {
-                        throw new Error(`Unknown command: ${arg}`);
+                        throw new Error(`Unknown command: ${arg}`)
                     }
                 } else {
-                    this.controller = controllers.find(x => matchName(x, arg));
+                    this.controller = controllers.find(x => matchName(x, arg))
                     if (!this.controller) {
-                        const byAlias = findControllerAndCommandByAlias(arg);
+                        const byAlias = findControllerAndCommandByAlias(arg)
                         if (byAlias) {
-                            this.controller = byAlias.controller;
-                            this.setCommand(byAlias.command);
+                            this.controller = byAlias.controller
+                            this.setCommand(byAlias.command)
                         } else if (arg.toLowerCase().trim() === "info") {
-                            this.printSummaryInfo = true;
-                            break;
+                            this.printSummaryInfo = true
+                            break
                         } else {
-                            throw new Error(`Unknown tool: ${arg}.`);
+                            throw new Error(`Unknown tool: ${arg}.`)
                         }
                     }
                 }
             }
         }
         if (this.pending) {
-            await this.resolveValue(this.pending, undefined);
+            await this.resolveValue(this.pending, undefined)
         }
         if (this.args.help || this.printSummaryInfo) {
-            return;
+            return
         }
         for (const arg of this.unresolved.values()) {
-            await this.resolveDefault(arg);
+            await this.resolveDefault(arg)
         }
     }
 }
 
 export async function missingArgError(arg: CommandArg): Promise<Error> {
-    const variants = await getArgVariants(arg);
+    const variants = await getArgVariants(arg)
     const variantsString = variants
         ? "\n" +
-        formatTable([
-                ["Available variants:", ""],
-                ...variants.map(x => [x.value, x.description ?? ""]),
-            ],
-            { headerSeparator: true },
-        )
-        : "";
-    throw new Error(`Missing required ${arg.name}${variantsString}`);
+          formatTable(
+              [
+                  ["Available variants:", ""],
+                  ...variants.map(x => [x.value, x.description ?? ""]),
+              ],
+              { headerSeparator: true },
+          )
+        : ""
+    throw new Error(`Missing required ${arg.name}${variantsString}`)
 }
 
 function isPrintVersionMode(): boolean {
     if (process.argv.length !== 3) {
-        return false;
+        return false
     }
-    const opt = process.argv[2].toLowerCase();
-    return opt === "--version" || opt === "-v";
+    const opt = process.argv[2].toLowerCase()
+    return opt === "--version" || opt === "-v"
 }
 
 export async function run(terminal: Terminal) {
-    const parser = new CommandLine();
-    const pkg = JSON.parse(fs.readFileSync(path.resolve(__dirname, "..", "..", "package.json"), "utf8"));
-   
+    const parser = new CommandLine()
+    const pkg = JSON.parse(
+        fs.readFileSync(
+            path.resolve(__dirname, "..", "..", "package.json"),
+            "utf8",
+        ),
+    )
+
     // Once a day, create a file containing the latest version number of `everdev`
     // Ignoring any (network, concurent access to file, etc) errors
-    createLatestVerFile(pkg.name).catch( _ => {} );
+    createLatestVerFile(pkg.name).catch(() => {})
 
-    const msg = getUpdateIsAvailableMsg(pkg.name, pkg.version);
+    const msg = getUpdateIsAvailableMsg(pkg.name, pkg.version)
     if (msg !== "") {
-        terminal.log(chalk.yellow(msg));
+        terminal.log(chalk.yellow(msg))
     }
 
     if (isPrintVersionMode()) {
-        terminal.log(pkg.version);
-        process.exit(0);
+        terminal.log(pkg.version)
+        return
     }
-    await parser.parse(process.argv.slice(2));
+    await parser.parse(process.argv.slice(2))
     if (parser.printSummaryInfo) {
-        await printSummaryInfo(terminal);
-        return;
+        await printSummaryInfo(terminal)
+        return
     }
-    const {
-        controller,
-        command,
-        args,
-    } = parser;
+    const { controller, command, args } = parser
+
     if (!controller || !command) {
-        await printUsage(terminal, controller, command);
-        return;
+        await printUsage(terminal, controller, command)
+        return
     }
     if (parser.args.help) {
-        await printUsage(terminal, controller, command);
-        return;
+        await printUsage(terminal, controller, command)
+        return
     }
-    await command.run(terminal, args);
+    await command.run(terminal, args)
 }
