@@ -1,3 +1,4 @@
-export {controllers, runCommand} from "./controllers/index";
+export { controllers, runCommand } from "./controllers/index"
 export * from "./core/utils"
 export * from "./core/index"
+export * as cli from "./cli/index"
